refactor: Move modules into seperate folder
Moved all movules files into their own folder with an index.js

undefined
