+++ lib/client.js
@@ -1,7 +1,5 @@
 const {Client} = require('discord.js')
-const {EventModule} = require('./event-module')
-const {PrefixModule} = require('./prefix-module')
-const {CommandModule} = require('./command-module')
+const {EventModule, PrefixModule, CommandModule} = require('./modules')
 
 /**
  * Discord.js Client with framework built on top
