+++ lib/modules/command.js
@@ -1,9 +1,9 @@
 const {Collection} = require('discord.js')
 const {Module} = require('./module')
-const {GroupModule} = require('./group-module')
-const defaultCommands = require('./commands')
-const {defaultNames} = require('./util/default-names')
-const {lowercase} = require('./util/lowercase')
+const {GroupModule} = require('./group')
+const defaultCommands = require('../commands')
+const {defaultNames} = require('../util/default-names')
+const {lowercase} = require('../util/lowercase')
 
 /**
  * Module for handling commands
