+++ lib/modules/event.js
@@ -1,6 +1,6 @@
 const {Module} = require('./module')
-const defaultEvents = require('./events')
-const {defaultNames} = require('./util/default-names')
+const defaultEvents = require('../events')
+const {defaultNames} = require('../util/default-names')
 
 /**
  * Module for adding events
