+++ lib/modules/group.js
@@ -1,8 +1,8 @@
 const {Collection} = require('discord.js')
 const {Module} = require('./module')
-const defaultGroups = require('./groups')
-const {defaultNames} = require('./util/default-names')
-const {lowercase} = require('./util/lowercase')
+const defaultGroups = require('../groups')
+const {defaultNames} = require('../util/default-names')
+const {lowercase} = require('../util/lowercase')
 
 /**
  * Module for handling command groups
