+++ lib/modules/index.js
@@ -1,0 +1,13 @@
+const {Module} = require('./module')
+const {EventModule} = require('./event')
+const {CommandModule} = require('./command')
+const {GroupModule} = require('./group')
+const {PrefixModule} = require('./prefix')
+
+module.exports = {
+  Module,
+  EventModule,
+  CommandModule,
+  GroupModule,
+  PrefixModule
+}
