+++ lib/modules/module.js
